+++ src/lib.rs
@@ -1,9 +1,13 @@
 //! Concise trait implementations.
 
+#![no_std]
 #![forbid(unsafe_code)]
 #![deny(rust_2018_idioms, nonstandard_style)]
 #![warn(future_incompatible)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
+
+#[cfg(test)]
+extern crate alloc;
 
 #[macro_use]
 mod macros;